Move even more stuff to internal/.
